Only GnuPG supports multiple private keys per user; restrict PrivateKey
As SMIME support requires that the private key for a user be in a .pem
file named with the email address, there can be only one private key for
a user at a time.  As such, PrivateKey (which also expects a ->{Key} on
key informational hashes, which SMIME lacks) and SetPrivateKey are
meaningless.