indexes on cached group members
CachedGroupMembers (CGM) table is mostly pass through
table when we either have GroupId and need MemberId or
have MemberId and need GroupId.

We also filter by Disabled column in most cases and don't
in some, but selectivity of Disabled column is very small,
so we put it as last.

Some DBs (mysql, newer Pg, most probably Oracle) can use
index to complete select without looking into table itself,
so we create index for both case mentioned above with Disabled
on the tail.

Also, we make sure (MemberId, ImmediateParentId) is indexed
which is used when we delete members from Groups.