Fix which part is labelled "Top" in signature attachments
VerifyDecrypt strips the signature part out of its parent, and to do
this it must be passed "Top".  However, during detection of detached
signatures, Top is set to the parent of the data part.  In almost all
cases, this is the same as the parent of the signature part; however,
the more correct solution is to set it to the parent of the signature
part, guaranteeing we will later find the signature to remove it.