Remove AddStatus/SetStatus arguments to VerifyDecrypt
There are two callsites of VerifyDecrypt: in
RT::Interface::Email::Auth::GnuPG, and in
share/html/Elements/ShowGnuPGStatus.  The latter omits AddStatus,
defaulting to SetStatus -- but this is both likely an erroneous omission
and irrelevant, as it ignores the modified $entity.  AddStatus is also
the always the more correct form, as parts may be multiply signed.