Allow safe_run_child to run before ConnectToDatabase runs
safe_run_child is primarily concerned with ensuring the database
connection is not lost or passon on to the child by accident.  However,
it should not error if the database connection does not exist yet, such
as during configuration loading.

Bracket all examination of $RT::Handle with guards to ensure that it is
defined.